% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CorrPlotTotal}
\alias{CorrPlotTotal}
\title{correlation plot}
\usage{
CorrPlotTotal(
  data1 = data1,
  data2 = data2,
  type1 = NULL,
  type2 = NULL,
  method = "spearman",
  topn1 = 50,
  topn2 = 50,
  cutoff_cor = 0.6,
  cutoff_p = 0.05
)
}
\value{
Correlation analysis of Multi-omics data.
}
\description{
tow omics correlation
}
\details{
Input two omics data matrix
}
