% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{CorrPlotTotal}
\alias{CorrPlotTotal}
\title{correlation plot}
\usage{
CorrPlotTotal(
  data1 = data1,
  data2 = data2,
  type1 = "omics1",
  type2 = "omics2",
  method = "spearman",
  topn1 = 50,
  topn2 = 50,
  cutoff_cor = 0.6,
  cutoff_p = 0.05
)
}
\arguments{
\item{data1}{one omics matrix}

\item{data2}{another omics matrix}

\item{type1}{data1 omics type}

\item{type2}{data2 omics type}

\item{method}{correlation method one of "spearman","person",defult is "spearman"}
}
\value{
Correlation analysis of Multi-omics data.
}
\description{
tow omics correlation
}
\details{
Input two omics data matrix
}
